requirements.txt
Pillow==9.5.0
pillow-heif==0.13.0
psd-tools==1.9.24
